Refatorando alguns detalhes para tornar o projeto mais legível
