+++ src/main/java/com/mieker/ifpr/shelfie/controller/AdminController.java
@@ -1,6 +1,6 @@
 package com.mieker.ifpr.shelfie.controller;
 
-import com.mieker.ifpr.shelfie.dto.RegisterUserDTO;
+import com.mieker.ifpr.shelfie.dto.User.RegisterUserDTO;
 import com.mieker.ifpr.shelfie.entity.User;
 import com.mieker.ifpr.shelfie.service.UserService;
 import org.springframework.http.ResponseEntity;
