+++ src/main/java/com/mieker/ifpr/shelfie/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.mieker.ifpr.shelfie.controller;
 
-import com.mieker.ifpr.shelfie.dto.UpdateUserDTO;
-import com.mieker.ifpr.shelfie.dto.UserDTO;
+import com.mieker.ifpr.shelfie.dto.User.UpdateUserDTO;
+import com.mieker.ifpr.shelfie.dto.User.UserDTO;
 import com.mieker.ifpr.shelfie.entity.User;
 import com.mieker.ifpr.shelfie.exception.GlobalExceptionHandler;
 import com.mieker.ifpr.shelfie.service.UserService;
