+++ src/main/java/com/mieker/ifpr/shelfie/dto/MyBooks/MyBooksDTO.java
@@ -1,4 +1,4 @@
-package com.mieker.ifpr.shelfie.dto;
+package com.mieker.ifpr.shelfie.dto.MyBooks;
 
 import com.mieker.ifpr.shelfie.entity.enumeration.BookStatus;
 import lombok.Data;
@@ -6,7 +6,10 @@ import lombok.Data;
 import java.util.UUID;
 
 @Data
-public class UpdateMyBooksDTO {
+public class MyBooksDTO {
     private UUID id;
+    private UUID userId;
+    private UUID bookId;
     private BookStatus bookStatus;
+    private boolean enabled;
 }
