+++ src/main/java/com/mieker/ifpr/shelfie/dto/MyBooks/UpdateMyBooksDTO.java
@@ -1,4 +1,4 @@
-package com.mieker.ifpr.shelfie.dto;
+package com.mieker.ifpr.shelfie.dto.MyBooks;
 
 import com.mieker.ifpr.shelfie.entity.enumeration.BookStatus;
 import lombok.Data;
@@ -6,10 +6,7 @@ import lombok.Data;
 import java.util.UUID;
 
 @Data
-public class MyBooksDTO {
+public class UpdateMyBooksDTO {
     private UUID id;
-    private UUID userId;
-    private UUID bookId;
     private BookStatus bookStatus;
-    private boolean enabled;
 }
