+++ src/main/java/com/mieker/ifpr/shelfie/service/AuthenticationService.java
@@ -1,7 +1,7 @@
 package com.mieker.ifpr.shelfie.service;
 
-import com.mieker.ifpr.shelfie.dto.LoginDTO;
-import com.mieker.ifpr.shelfie.dto.RegisterUserDTO;
+import com.mieker.ifpr.shelfie.dto.User.LoginDTO;
+import com.mieker.ifpr.shelfie.dto.User.RegisterUserDTO;
 import com.mieker.ifpr.shelfie.entity.User;
 import com.mieker.ifpr.shelfie.entity.enumeration.UserRoles;
 import com.mieker.ifpr.shelfie.mapper.UserMapper;
