+++ src/main/java/com/mieker/ifpr/shelfie/service/UserService.java
@@ -1,8 +1,8 @@
 package com.mieker.ifpr.shelfie.service;
 
-import com.mieker.ifpr.shelfie.dto.RegisterUserDTO;
-import com.mieker.ifpr.shelfie.dto.UpdateUserDTO;
-import com.mieker.ifpr.shelfie.dto.UserDTO;
+import com.mieker.ifpr.shelfie.dto.User.RegisterUserDTO;
+import com.mieker.ifpr.shelfie.dto.User.UpdateUserDTO;
+import com.mieker.ifpr.shelfie.dto.User.UserDTO;
 import com.mieker.ifpr.shelfie.entity.User;
 import com.mieker.ifpr.shelfie.entity.enumeration.UserRoles;
 import com.mieker.ifpr.shelfie.exception.GlobalExceptionHandler;
@@ -15,7 +15,6 @@ import org.springframework.stereotype.Service;
 
 import java.text.ParseException;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 // user controller -> service -> repository -> dto
